* added types to its own file * tweaked cover_images * song arist, title, date moved to SongInfo
